Various improvements (#45)
* add comments

* use u128 for timestamp

* check non-zero amount when creating Coin/Coins

* add comments

* add try from iterator methods

* update usages

* use NonZero instead

* add message to test suite

* add methods to convert strings to Binary

* methods for TestSuite<RustVm> that don't need gas

* recompile tester contracts

+++ crates/app/src/abci.rs
@@ -164,7 +164,7 @@ fn from_tm_block(height: i64, time: Option<TmTimestamp>, hash: Option<Bytes>) ->
 }
 
 fn from_tm_timestamp(time: TmTimestamp) -> Timestamp {
-    Timestamp::from_seconds(time.seconds as u64).plus_nanos(time.nanos as u64)
+    Timestamp::from_seconds(time.seconds as u128).plus_nanos(time.nanos as u128)
 }
 
 fn from_tm_hash(bytes: Bytes) -> Hash {
